ユーザー登録完了後にログイン画面に遷移できるようにする
