+++ frontend/src/main.ts
@@ -1,4 +1,5 @@
 import './assets/main.css'
+import 'bootstrap/dist/css/bootstrap.min.css'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
